Added the concept of asynchronous actions.
Asynchronous actions represent tasks that must be performed by VISA
devices while not returning any values on successful completion (such
as a device reset). Defining an asynchronous action requires a method
of the AsyncAction delegate type that is additionally decorated with
the AsyncActionAttribute. Like asynchronous properties they can be
enumerated.