Fixed more asynchronous disconnection issues.
The normal device disconnection in the device control panel was
causing errors when any asynchronous actions or updates were running
in the background. Now all asynchronous actions are awaited for
completion during the device disconnection to prevent their failure
because of the sudden VISA session loss.
The manual update of all asynchronous properties is now executed the
same way as other device-related asynchronous actions, so it also will
be tracked at device disconnection.