Extracted the view model logic to the VisaDeviceBuilder library.
The main view model logic has been moved to the new VisaDeviceController
class in the VisaDeviceBuilder library. This makes it cross-platform as
it does not depend on the WPF framework any more. The WPF-dependant
features like commands remain in the previous view-model class that now
derives from the controller class.