Added the owned asynchronous property and device action concepts.
The owned asynchronous property is an ordinary asynchronous property,
but it also has an owner, so that getter and setter accessors can be
evaluated against this particular VISA device defined through the Owner
class property. The same logic is implemented for the owned device
action as well.
These concepts allow to detach delegates from a device they were
designed to be called for. They also are essential when there is a need
of creating multiple similar VISA devices using builders.