Updated the VisaDeviceBuilder logic.
Removed the generic VisaDeviceBuilder class, so that simple
VisaDeviceBuilder and MessageDeviceBuilder classes do not have any
common ancestors (except the generic IVisaDeviceBuilder interface),
and each has its own collection of builder methods.
This allows to use builder methods in any order without a need to
remember builder class inheritance. Earlier it was impossible to use the
UseMessageProcessor method in a chain after any other base method was
called because it was returning an ancestor builder type.
Nevertheless, it requires to make a full copy of unit tests covering the
VisaDeviceBuilder class to also cover the MessageDeviceBuilder class.