Reworked the OwnedAsyncProperty and OwnedDeviceAction concepts.
The logic of owned asynchronous property and device action variants
has been transferred to the original non-owned AsyncProperty and
DeviceAction classes. The Owner property has been also renamed to
TargetDevice and now it can be null by concept. This eliminates the
need of distinguishing owned and non-owned variants and makes the code
easier to maintain.
Also updated the related unit tests.