Added the VisaResourceName record.
This record is now returned in results of methods from the
VisaResourceLocator static class and stores both canonical and alias
names of found VISA resources. It allows to clean up the results where
previously canonical and alias names were mixed with each other.
It also can be implicitly cast into a string with the corresponding
canonical resource name, so that it can be safely assigned to the
IVisaDevice.ResourceName property as it is.