Reworked the device actions.
The DeviceAction class now does not depend on the DeviceActionExecutor
static class and is able to execute the wrapped action and track its
state by itself. It was an important change because the original idea of
centralized device action tracking within the DeviceActionExecutor class
could have lead to situations where one VISA device depends on device
action execution states of another VISA device which is wrong.
Also now an additional custom delegate can be passed into the
constructor that determines if the device action can be executed at the
moment.